Refactor indexer file names and indexer sample to snake case :snake:
